+++ src/main/java/com/owczarczak/footballers/contract/ContractDto.java
@@ -8,10 +8,10 @@ import java.time.LocalDate;
 @Data
 @Builder
 public class ContractDto {
-    private int id;
+    private Long id;
     private String clubName;
     private String footballerName;
     private LocalDate contractStart;
     private LocalDate contractEnd;
-    private int salary;
+    private Integer salary;
 }
